package/.ipynb_checkpoints/requirements-checkpoint.txt
tqdm
pyyaml
# scikit-learn==1.6.1
# pandas==1.5.3
# numpy==1.26.4

# pandas==2.0.3
# numpy
# joblib==1.3.2